Refactor: Move libstbt.so wrappers into libstbt.py
...and out of `sqdiff.py`.  `sqdiff.py` can now be imported even if
libstbt.so is not built - such as when installing the Python package
on Windows.
